backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Cors==4.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
PyJWT==2.8.0
google-generativeai>=0.8.0
requests==2.32.3
gunicorn==22.0.0
SQLAlchemy==2.0.31
